refactor(prover-ray): move the ABI Agreement Test to verifier-ray
