+++ verifier-ray/codegen/abicheck/abi_agreement_test.go
@@ -1,8 +1,8 @@
-// Tests that keep prover-ray's encoder and verifier-ray in step. Two halves:
-// the layout numbers must agree (TestABIAgreement), and the image fixture
-// verifier-ray reads must match what the encoder currently produces
+// Package abicheck keeps verifier-ray in step with the prover-ray encoder it
+// pins. Two halves: the layout numbers must agree (TestABIAgreement), and the
+// image fixture the Zig tests read must match what the pinned encoder produces
 // (TestVerifierRayImageIsUpToDate).
-package proofserialization_test
+package abicheck
 
 import (
 	"os"
@@ -14,12 +14,12 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// proofABIPath is verifier-ray's pinned layout, the authority this package
+// proofABIPath is verifier-ray's pinned layout, the authority the encoder
 // encodes against.
-const proofABIPath = "../../../verifier-ray/src/proof_abi.zig"
+const proofABIPath = "../../src/proof_abi.zig"
 
-// TestABIAgreement checks this package's size and offset constants against the
-// assertions in verifier-ray/src/proof_abi.zig.
+// TestABIAgreement checks the pinned encoder's size and offset constants against
+// the assertions in src/proof_abi.zig.
 //
 // This closes the one drift direction nothing else covers. proof_abi.zig catches
 // Zig's layout moving out from under the pins; the encoder's own tests catch Go
@@ -32,12 +32,9 @@ const proofABIPath = "../../../verifier-ray/src/proof_abi.zig"
 // not couple the Go tests to a Zig toolchain.
 func TestABIAgreement(t *testing.T) {
 	src, err := os.ReadFile(proofABIPath)
-	if err != nil {
-		t.Skipf("verifier-ray not checked out alongside prover-ray (%v); "+
-			"the ABI cross-check needs %s", err, proofABIPath)
-	}
+	require.NoError(t, err, "reading %s", proofABIPath)
 
-	// wantSize maps a Zig type as written in proof_abi.zig to this package's
+	// wantSize maps a Zig type as written in proof_abi.zig to the encoder's
 	// corresponding size constant. Types with no Go counterpart in the image
 	// (the slice header itself is checked via SizeSlice) are listed explicitly so
 	// a newly pinned type shows up as unmapped rather than being ignored.
@@ -61,7 +58,7 @@ func TestABIAgreement(t *testing.T) {
 		"?merkle.RowPair":         ps.SizeOptRowPair,
 	}
 
-	// wantOffset maps a pinned (type, field) to this package's offset constant.
+	// wantOffset maps a pinned (type, field) to the encoder's offset constant.
 	wantOffset := map[[2]string]int{
 		{"ext.Ext", "B0"}: 0,
 		{"ext.Ext", "B1"}: 8,
@@ -96,7 +93,7 @@ func TestABIAgreement(t *testing.T) {
 		{"verifier.Proof", "pcs_opening"}:  ps.OffProofPcsOpening,
 	}
 
-	// These must reference the package's constants, not literal copies of the
+	// These must reference the encoder's constants, not literal copies of the
 	// pinned values. Hardcoding the numbers here compares Zig's pin against a
 	// copy of itself, which passes no matter what the encoder actually writes —
 	// a mutation run caught exactly that, with TagColumnPublic changed to 2 and
@@ -116,7 +113,7 @@ func TestABIAgreement(t *testing.T) {
 	for _, m := range sizes {
 		zigType, want := m[1], mustAtoi(t, m[2])
 		got, ok := wantSize[zigType]
-		require.True(t, ok, "%s pins a size for %q that this package does not map to a "+
+		require.True(t, ok, "%s pins a size for %q that this check does not map to a "+
 			"constant; add it to wantSize (and to the encoder if the image carries it)",
 			proofABIPath, zigType)
 		require.Equal(t, want, got,
@@ -130,7 +127,7 @@ func TestABIAgreement(t *testing.T) {
 		key := [2]string{m[1], m[2]}
 		want := mustAtoi(t, m[3])
 		got, ok := wantOffset[key]
-		require.True(t, ok, "%s pins an offset for %s.%s that this package does not map; "+
+		require.True(t, ok, "%s pins an offset for %s.%s that this check does not map; "+
 			"add it to wantOffset", proofABIPath, key[0], key[1])
 		require.Equal(t, want, got,
 			"offset of %s.%s: %s pins %d, proofserialization uses %d — the encoder would write "+
@@ -143,7 +140,7 @@ func TestABIAgreement(t *testing.T) {
 		key := [2]string{m[1], m[2]}
 		want := mustAtoi(t, m[3])
 		got, ok := wantTag[key]
-		require.True(t, ok, "%s pins a discriminant for %s.%s that this package does not map; "+
+		require.True(t, ok, "%s pins a discriminant for %s.%s that this check does not map; "+
 			"add it to wantTag", proofABIPath, key[0], key[1])
 		require.Equal(t, want, got,
 			"discriminant of %s.%s: %s pins %d, proofserialization uses %d — the encoder would "+
@@ -158,10 +155,10 @@ func mustAtoi(t *testing.T, s string) int {
 	return n
 }
 
-// verifierRayImagePath is the image verifier-ray's proof_image_test.zig maps and
-// reads. It is the only test in which a byte produced by this encoder is
-// consumed by the actual verifier rather than by this package's own decoder.
-const verifierRayImagePath = "../../../verifier-ray/testdata/proof_image.bin"
+// verifierRayImagePath is the image test/proof_image_test.zig maps and reads.
+// It is the only test in which a byte produced by the encoder is consumed by the
+// actual verifier rather than by the encoder's own decoder.
+const verifierRayImagePath = "../../testdata/proof_image.bin"
 
 // verifierRayImageBase is the address the fixture image is relocated for, and
 // the address verifier-ray's Zig test maps it at with MAP_FIXED.
@@ -236,9 +233,9 @@ func verifierRayFixture() ps.VerifyInput {
 
 // TestVerifierRayImageIsUpToDate keeps the committed cross-language fixture in sync.
 //
-// The image is committed rather than generated at Zig test time so verifier-ray's
-// suite stays self-contained — it needs no Go toolchain. The cost is that the
-// file can go stale, which is what this test prevents. Regenerate with
+// The image is committed rather than generated at Zig test time so the Zig suite
+// stays self-contained — it needs no Go toolchain. The cost is that the file can
+// go stale, which is what this test prevents. Regenerate with
 // UPDATE_VERIFIER_RAY_IMAGE=1.
 func TestVerifierRayImageIsUpToDate(t *testing.T) {
 	image, err := ps.Encode(verifierRayFixture(), verifierRayImageBase)
@@ -260,14 +257,11 @@ func TestVerifierRayImageIsUpToDate(t *testing.T) {
 	}
 
 	committed, err := os.ReadFile(verifierRayImagePath)
-	if err != nil {
-		t.Skipf("verifier-ray not checked out alongside prover-ray (%v); "+
-			"run UPDATE_VERIFIER_RAY_IMAGE=1 go test ./wiop/proofserialization/ to create %s",
-			err, verifierRayImagePath)
-	}
+	require.NoError(t, err, "reading %s", verifierRayImagePath)
 
 	require.Equal(t, committed, image,
-		"the image verifier-ray reads is stale. verifier-ray's proof_image_test.zig "+
-			"asserts against it, so regenerate with UPDATE_VERIFIER_RAY_IMAGE=1 and re-run "+
-			"`zig build test` in verifier-ray to confirm the Zig side still agrees")
+		"the image verifier-ray reads is stale against the pinned prover-ray. "+
+			"test/proof_image_test.zig asserts against it, so regenerate with "+
+			"UPDATE_VERIFIER_RAY_IMAGE=1 go test ./abicheck/ and re-run `zig build test` "+
+			"to confirm the Zig side still agrees")
 }
